+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changes
 - refactorings #3373
+- delete outgoing MDNs found in the Sent folder on Gmail #3372
 
 ## 1.84.0